Add GradCAM for object detection (#75)
* Add example with GradCAM for YOLOv5 in object detection

* Remove constraints for min and max bbox size

* Extract base ObjectDetector class

* Remove previous example with object detection task

* Add original code source URL in docstring

* Move fetching layer from YOLO model function, add target_layer parameter to GradCAM

* Add dataclasses for complex object detection output data types

* Refactor code, add docstrings, fix names, add workaround with np.abs for bbox coordinates

* Add support for torchvision SSD model for object detection, refactoring

* Rename directories

* Refactor, change BaseObjectDetector class definition, make SSD model inherit from it

* Refactor modules structure, add GradCAM to object detection module

* Remove unused imports

* Refactor, move OD models to separate module, fix YOLO prediction generation

* Apply pre-commit hooks to all files

* Replace excessive unnecessary dependency with simple parsing

* Remove model warmup run

* Remove unused path in forward pass algorithm

* Replace custom implementations with torchvision imports

* Add unit tests for object detection utils

* Remove obsolete directory

* Fix YOLOv5 bbox conversion

* Add unit test for object detection visualization utils, refactor

* Refactor GradCAM for OD

* Remove redundant device argument to YOLOv5ObjectDetector class initializer

* Fix unit tests for image preprocessing - use rectangle instead of square shapes

* Simplify forward function in WrapperYOLOv5ObjectDetectionModule class

* Add custom GradCAM base algorithm implementation for classification and object detection

* Move object detection model examples to examples directory

* Restructure library directory structure for explainer algorithms, move object detection example script to notebook

* Add interpolation method parameter to resize_image function

* Fix unit tests and imports

* Remove adding epsilon in preprocess object detection image function

* Fix example notebook - add assertions of YOLOv5 image shape

* Fix basic usage notebook after refactoring class names and directory structure

* Enable changing image ratio to match YOLOv5 requirements of image shape

* Update README

* Refactor object detection custom modules

* Refactor GradCAM for object detection

+++ foxai/callbacks/wandb_callback.py
@@ -6,12 +6,12 @@ import matplotlib
 import numpy as np
 import pytorch_lightning as pl
 import torch
+import wandb
 from pytorch_lightning.loggers import WandbLogger
 from torch.utils.data import DataLoader
 
-import wandb
 from foxai.array_utils import convert_standardized_float_to_uint8, standardize_array
-from foxai.context_manager import FoXaiExplainer, ExplainerWithParams
+from foxai.context_manager import ExplainerWithParams, FoXaiExplainer
 from foxai.visualizer import mean_channels_visualization
 
 AttributeMapType = Dict[str, List[np.ndarray]]
@@ -118,7 +118,7 @@ class WandBCallback(pl.callbacks.Callback):
             )
             figures_dict[explainer_name].append(figure)
             standardized_attr = standardize_array(
-                explainer_attributes.detach().cpu().numpy()
+                explainer_attributes.detach().cpu().numpy().astype(float)
             )
 
             attributes_dict[explainer_name].append(
